refactor: split route handlers into routes/ directory (#66)
Extract all FastAPI route functions from api.py into a routes/ package:
- routes/markets.py — market CRUD, comments, resolution committee, history
- routes/trading.py — bet, sell, positions, bet history
- routes/agents.py — registration, profiles, reputation, leaderboard, claim
- routes/chat.py — send and retrieve chat messages
- routes/admin.py — admin user management
- routes/meta.py — health, currency, skill.md

api.py retains: app setup, middleware, CORS, auth deps, helpers, constants,
db instance, economics config, payout helper, and _SKILL_MD template.

Route registration uses deferred imports inside _register_routes() to avoid
circular imports (route modules import shared symbols from api.py).

No behavioral changes — all 36 routes register identically.
