+++ api.py
@@ -440,1184 +440,11 @@ def _calculate_and_distribute_payouts(market_id: str, outcome: Outcome) -> int:
     return paid
 
 
-# =============================================================================
-# Market Endpoints
-# =============================================================================
-
-@app.get("/markets", response_model=List[MarketSummary], tags=["markets"])
-async def list_markets(
-    request: Request,
-    response: Response,
-    status: Optional[str] = None,
-):
-    """List markets, filtered by status.
-
-    Query params:
-        status: Filter by market status.
-            - omitted or "active" or "open" → only OPEN markets (default)
-            - "resolving"       → markets past closes_at, awaiting resolution
-            - "closed" or "resolved" → resolved markets
-            - "all"             → all markets regardless of status
-
-    Caching:
-        Responses include ETag and Last-Modified headers for HTTP caching.
-        Send `If-None-Match` or `If-Modified-Since` to receive 304 Not Modified
-        when data hasn't changed, saving bandwidth and parse time.
-        Server-side results are cached in-memory with a 5-second TTL and
-        invalidated immediately on any market mutation (create, bet, sell, resolve).
-    """
-    status_filter = (status or "active").strip().upper()
-
-    # ── HTTP conditional request: If-None-Match ──
-    # Check before doing any work — if the client already has the current version,
-    # return 304 immediately (no DB query, no serialization).
-    client_etag = request.headers.get("if-none-match")
-    cached = market_cache.get(status_filter)
-    if cached and client_etag and client_etag == cached["etag"]:
-        return Response(
-            status_code=304,
-            headers={
-                "ETag": cached["etag"],
-                "Last-Modified": market_cache.last_modified.strftime("%a, %d %b %Y %H:%M:%S GMT"),
-                "Cache-Control": "public, max-age=5",
-            },
-        )
-
-    # ── In-memory cache hit ──
-    if cached:
-        _set_cache_headers(response, cached["etag"], market_cache.last_modified)
-        return cached["data"]
-
-    # ── Cache miss: build response from DB ──
-    # Single JOIN query: markets + creator usernames (was N+1: 1 + N get_user calls)
-    markets = db.list_markets_with_creators()
-
-    # Auto-transition: move OPEN markets past closes_at to RESOLVING
-    now = datetime.now(timezone.utc)
-    transitioned = False
-    for m in markets:
-        if m["status"] == MarketStatus.OPEN and m["closes_at"] <= now:
-            db.update_market_status(m["id"], MarketStatus.RESOLVING)
-            m["status"] = MarketStatus.RESOLVING
-            transitioned = True
-
-    # If we transitioned any markets, invalidate cache so other status filters
-    # pick up the change too.
-    if transitioned:
-        market_cache.invalidate()
-
-    # Apply status filter (default: only open/active markets)
-    if status_filter == "ALL":
-        pass  # no filtering
-    elif status_filter in ("CLOSED", "RESOLVED"):
-        markets = [m for m in markets if m["status"] == MarketStatus.RESOLVED]
-    elif status_filter == "RESOLVING":
-        markets = [m for m in markets if m["status"] == MarketStatus.RESOLVING]
-    else:
-        # Default: only open markets (ACTIVE or OPEN both map here)
-        markets = [m for m in markets if m["status"] == MarketStatus.OPEN]
-
-    # Build response — precompute probability once per market
-    result = []
-    for m in markets:
-        result.append(MarketSummary(
-            id=m["id"],
-            title=m["title"],
-            probability=get_cpmm_probability(m["pool"], m["p"]),
-            status=m["status"],
-            closes_at=m["closes_at"],
-            total_volume=m["total_volume"],
-            creator_id=m["creator_id"],
-            creator_username=m.get("creator_username"),
-        ))
-
-    # Store in cache
-    entry = market_cache.set(status_filter, result)
-    _set_cache_headers(response, entry["etag"], market_cache.last_modified)
-    return result
-
-
-def _set_cache_headers(response: Response, etag: str, last_modified: datetime) -> None:
+def _set_cache_headers(response, etag: str, last_modified) -> None:
     """Set ETag, Last-Modified, and Cache-Control headers on a response."""
     response.headers["ETag"] = etag
     response.headers["Last-Modified"] = last_modified.strftime("%a, %d %b %Y %H:%M:%S GMT")
     response.headers["Cache-Control"] = "public, max-age=5"
-
-
-@app.get("/markets/{market_id}", response_model=MarketDetail, tags=["markets"])
-async def get_market(market_id: str):
-    """Get market details including current probability."""
-    _validate_uuid(market_id, "market_id")
-    market = db.get_market(market_id)
-    if not market:
-        return error_response(404, "Market not found", ErrorCode.MARKET_NOT_FOUND)
-    
-    # Auto-transition: OPEN → RESOLVING when closes_at has passed
-    now = datetime.now(timezone.utc)
-    if market["status"] == MarketStatus.OPEN and market["closes_at"] <= now:
-        db.update_market_status(market_id, MarketStatus.RESOLVING)
-        market["status"] = MarketStatus.RESOLVING
-    
-    # Look up creator username
-    creator = db.get_user(market["creator_id"]) if market["creator_id"] else None
-    creator_username = creator["username"] if creator else None
-    
-    return MarketDetail(
-        id=market["id"],
-        title=market["title"],
-        description=market["description"],
-        probability=get_cpmm_probability(market["pool"], market["p"]),
-        status=market["status"],
-        closes_at=market["closes_at"],
-        created_at=market["created_at"],
-        resolved_at=market["resolved_at"],
-        resolution=market["resolution"],
-        total_volume=market["total_volume"],
-        creator_id=market["creator_id"],
-        creator_username=creator_username,
-        pool=market["pool"],
-        p=market["p"],
-    )
-
-
-@app.post("/markets", response_model=MarketCreated, tags=["markets"])
-async def create_market(req: MarketCreate, user: dict = Depends(require_auth)):
-    """Create a new prediction market."""
-    # Require twitter verification before creating markets
-    if user.get("status") != "claimed":
-        return error_response(403,
-            "Twitter verification required before creating markets. Visit /claim/{user_id} to link your Twitter account.",
-            ErrorCode.CLAIM_REQUIRED)
-    
-    now = datetime.now(timezone.utc)
-
-    if req.closes_at <= now:
-        return error_response(400, "closes_at must be in the future", ErrorCode.INVALID_INPUT)
-
-    # Enforce max market duration (testing phase)
-    max_close = now + timedelta(seconds=MAX_MARKET_DURATION_SECONDS)
-    if req.closes_at > max_close:
-        return error_response(422,
-            "Market duration cannot exceed 1 hour during testing phase",
-            ErrorCode.MARKET_DURATION_EXCEEDED)
-
-    # Check creator has enough balance to fund the initial liquidity pool
-    if user["balance"] < MARKET_CREATION_COST:
-        return error_response(400,
-            f"Insufficient balance. Market creation costs {MARKET_CREATION_COST}{CURRENCY_SYMBOL}.",
-            ErrorCode.INSUFFICIENT_BALANCE,
-            detail={"balance": user["balance"], "required": MARKET_CREATION_COST})
-
-    # Rate limit check: cabal members get 1-min cooldown, everyone else 30-min
-    username = user.get("username", "").lower()
-    cooldown_minutes = CABAL_COOLDOWN_MINUTES if username in CABAL_USERNAMES else DEFAULT_COOLDOWN_MINUTES
-    last_created = user.get("last_market_created_at")
-    if last_created:
-        # Handle both datetime objects and strings
-        if isinstance(last_created, str):
-            last_created = datetime.fromisoformat(last_created.replace('Z', '+00:00'))
-        cooldown_end = last_created + timedelta(minutes=cooldown_minutes)
-        if now < cooldown_end:
-            remaining = (cooldown_end - now).total_seconds() / 60
-            return error_response(429,
-                f"Rate limit: you can create another market in {remaining:.0f} minutes",
-                ErrorCode.RATE_LIMITED,
-                detail={"retry_after_minutes": round(remaining, 1)})
-    
-    # Deduct creation cost from creator's balance (funds the initial liquidity pool)
-    db.update_user_balance(user["id"], -MARKET_CREATION_COST)
-
-    market_id = str(uuid.uuid4())
-    market = db.create_market(
-        market_id=market_id,
-        creator_id=user["id"],
-        title=req.title,
-        description=req.description,
-        closes_at=req.closes_at,
-        initial_liquidity=req.initial_liquidity,
-    )
-    
-    # Update last market creation timestamp
-    db.update_user_last_market_created(user["id"])
-    
-    # Invalidate market list cache — new market should appear immediately
-    market_cache.invalidate()
-    
-    # Calculate market duration for guidance
-    now = datetime.now(timezone.utc)
-    duration_days = (req.closes_at - now).total_seconds() / 86400
-    
-    # Generate guidance based on duration
-    tip = None
-    warning = None
-    
-    if duration_days <= 7:
-        tip = "Nice! Short markets (under 7 days) typically see 2-3x more trading activity."
-    elif duration_days > 14:
-        warning = "Heads up: markets over 2 weeks often see lower engagement. Consider shorter timeframes for more action."
-    
-    return MarketCreated(
-        id=market["id"],
-        title=market["title"],
-        description=market["description"],
-        probability=get_cpmm_probability(market["pool"], market["p"]),
-        status=market["status"],
-        closes_at=market["closes_at"],
-        created_at=market["created_at"],
-        resolved_at=market["resolved_at"],
-        resolution=market["resolution"],
-        total_volume=market["total_volume"],
-        creator_id=market["creator_id"],
-        creator_username=user["username"],
-        pool=market["pool"],
-        p=market["p"],
-        creation_cost=MARKET_CREATION_COST,
-        tip=tip,
-        warning=warning,
-    )
-
-
-@app.post("/markets/{market_id}/resolve", response_model=MarketDetail, tags=["markets"])
-async def resolve_market(market_id: str, req: MarketResolve, user: dict = Depends(require_auth)):
-    """Resolve a market. Only creator can resolve."""
-    _validate_uuid(market_id, "market_id")
-    market = db.get_market(market_id)
-    if not market:
-        return error_response(404, "Market not found", ErrorCode.MARKET_NOT_FOUND)
-    
-    if market["creator_id"] != user["id"]:
-        return error_response(403, "Only creator can resolve market", ErrorCode.FORBIDDEN)
-    
-    if market["status"] == MarketStatus.RESOLVED:
-        return error_response(400, "Market already resolved", ErrorCode.ALREADY_RESOLVED)
-    
-    # Auto-transition OPEN → RESOLVING if closes_at has passed
-    now = datetime.now(timezone.utc)
-    if market["status"] == MarketStatus.OPEN and market["closes_at"] <= now:
-        db.update_market_status(market_id, MarketStatus.RESOLVING)
-        market["status"] = MarketStatus.RESOLVING
-    
-    db.resolve_market(market_id, req.outcome)
-    _calculate_and_distribute_payouts(market_id, req.outcome)
-    
-    # Invalidate market list cache — status changed to RESOLVED
-    market_cache.invalidate()
-    
-    market = db.get_market(market_id)
-    
-    # Look up creator username
-    creator = db.get_user(market["creator_id"]) if market["creator_id"] else None
-    creator_username = creator["username"] if creator else None
-    
-    return MarketDetail(
-        id=market["id"],
-        title=market["title"],
-        description=market["description"],
-        probability=1.0 if req.outcome == Outcome.YES else 0.0,
-        status=market["status"],
-        closes_at=market["closes_at"],
-        created_at=market["created_at"],
-        resolved_at=market["resolved_at"],
-        resolution=market["resolution"],
-        total_volume=market["total_volume"],
-        creator_id=market["creator_id"],
-        creator_username=creator_username,
-        pool=market["pool"],
-        p=market["p"],
-    )
-
-
-# =============================================================================
-# Trading Endpoints
-# =============================================================================
-
-@app.post("/markets/{market_id}/bet", response_model=BetResponse, tags=["trading"])
-async def place_bet(market_id: str, req: BetRequest, response: Response, user: dict = Depends(require_auth)):
-    """Place a bet on a market.
-    
-    Rate limited: max 30 bets per agent per minute.
-    Max bet amount: 500ŧ per single bet.
-    
-    Rate limit headers are included in every response:
-    - `X-RateLimit-Limit`: Max requests in window
-    - `X-RateLimit-Remaining`: Remaining requests
-    - `X-RateLimit-Reset`: Unix timestamp when window resets
-    - `Retry-After` (on 429 only): Seconds to wait
-    """
-    _validate_uuid(market_id, "market_id")
-    # Require twitter verification before trading
-    if user.get("status") != "claimed":
-        return error_response(403,
-            "Twitter verification required before trading. Visit /claim/{user_id} to link your Twitter account.",
-            ErrorCode.CLAIM_REQUIRED)
-    
-    # ── Max bet amount ──
-    if req.amount > MAX_BET_AMOUNT:
-        return error_response(400,
-            f"Bet amount {req.amount}ŧ exceeds maximum of {MAX_BET_AMOUNT}ŧ per bet.",
-            ErrorCode.INVALID_INPUT,
-            detail={"amount": req.amount, "max": MAX_BET_AMOUNT})
-
-    # ── Rate limit: bets per agent ──
-    allowed, info = rate_limiter.check(
-        f"bet:{user['id']}",
-        max_requests=MAX_BETS_PER_MINUTE,
-        window_seconds=60,
-    )
-    if not allowed:
-        raise_rate_limited(
-            f"Betting rate limit exceeded ({MAX_BETS_PER_MINUTE}/minute). {info['detail']}",
-            info,
-        )
-    # Inject rate limit headers on successful responses too
-    set_rate_limit_headers(response, info)
-
-    market = db.get_market(market_id)
-    if not market:
-        return error_response(404, "Market not found", ErrorCode.MARKET_NOT_FOUND)
-    
-    # Auto-transition: OPEN → RESOLVING when closes_at has passed
-    now = datetime.now(timezone.utc)
-    if market["status"] == MarketStatus.OPEN and market["closes_at"] <= now:
-        db.update_market_status(market_id, MarketStatus.RESOLVING)
-        market["status"] = MarketStatus.RESOLVING
-    
-    if market["status"] != MarketStatus.OPEN:
-        status_msg = "Market is resolving (closed, awaiting resolution)" if market["status"] == MarketStatus.RESOLVING else "Market is not open for trading"
-        return error_response(400, status_msg, ErrorCode.MARKET_CLOSED)
-    
-    if market["closes_at"] <= now:
-        return error_response(400, "Market has closed", ErrorCode.MARKET_CLOSED)
-    
-    # Calculate trade fee (2% total)
-    trade_fee = req.amount * TRADE_FEE_RATE
-    total_cost = req.amount + trade_fee
-    
-    if user["balance"] < total_cost:
-        return error_response(400,
-            f"Insufficient balance. Need {total_cost:.2f} (bet: {req.amount:.2f} + fee: {trade_fee:.2f})",
-            ErrorCode.INSUFFICIENT_BALANCE,
-            detail={"balance": user["balance"], "required": total_cost})
-    
-    # Calculate bet using CPMM
-    prob_before = get_cpmm_probability(market["pool"], market["p"])
-    
-    state = CpmmState(pool=market["pool"].copy(), p=market["p"])
-    result = calculate_cpmm_purchase(state, req.amount, req.outcome.value)
-    
-    shares = result["shares"]
-    if shares <= 0:
-        return error_response(400, "Trade would result in zero or negative shares", ErrorCode.ZERO_SHARES)
-    
-    prob_after = get_cpmm_probability(result["new_pool"], result["new_p"])
-    
-    # Execute trade - deduct bet amount + fee from user
-    db.update_user_balance(user["id"], -total_cost)
-    
-    # Pay creator their share of the fee (1%)
-    creator_fee = trade_fee * CREATOR_FEE_SHARE
-    if market["creator_id"] != user["id"]:  # Don't pay yourself
-        db.update_user_balance(market["creator_id"], creator_fee)
-    # Note: remaining fee (1%) is burned - not allocated to anyone
-    
-    db.update_market_pool(market_id, result["new_pool"], result["new_p"], req.amount)
-    db.update_position(market_id, user["id"], req.outcome, shares, req.amount)
-    
-    # Invalidate market list cache — probability and volume changed
-    market_cache.invalidate()
-    
-    bet_id = str(uuid.uuid4())
-    bet = db.create_bet(
-        bet_id=bet_id,
-        market_id=market_id,
-        user_id=user["id"],
-        outcome=req.outcome,
-        amount=req.amount,
-        shares=shares,
-        prob_before=prob_before,
-        prob_after=prob_after,
-    )
-    
-    # Fetch updated balance for the response
-    updated_user = db.get_user(user["id"])
-    new_balance = updated_user["balance"] if updated_user else user["balance"] - total_cost
-    
-    return BetResponse(
-        bet_id=bet["id"],
-        market_id=bet["market_id"],
-        market_title=market["title"],
-        user_id=bet["user_id"],
-        outcome=bet["outcome"],
-        amount=bet["amount"],
-        fee=trade_fee,
-        fee_breakdown=FeeBreakdown(
-            total_fee=trade_fee,
-            creator_fee=creator_fee,
-            platform_fee=trade_fee - creator_fee,
-        ),
-        total_cost=total_cost,
-        new_balance=round(new_balance, 8),
-        shares=bet["shares"],
-        avg_price=bet["avg_price"],
-        probability_before=bet["probability_before"],
-        probability_after=bet["probability_after"],
-        created_at=bet["created_at"],
-    )
-
-
-# Alias: accept POST /markets/{id}/bets (plural) — redirects to the singular handler
-# Some SDKs/clients expect the plural form. Both work identically.
-@app.post("/markets/{market_id}/bets", response_model=BetResponse, tags=["trading"])
-async def place_bet_plural_alias(market_id: str, req: BetRequest, user: dict = Depends(require_auth)):
-    """Place a bet on a market (alias for POST /markets/{id}/bet)."""
-    return await place_bet(market_id, req, user)
-
-
-@app.post("/markets/{market_id}/sell", response_model=SellResponse, tags=["trading"])
-async def sell_shares(market_id: str, req: SellRequest, user: dict = Depends(require_auth)):
-    """Sell shares back to the market."""
-    _validate_uuid(market_id, "market_id")
-    # Require twitter verification before trading
-    if user.get("status") != "claimed":
-        return error_response(403,
-            "Twitter verification required before trading. Visit /claim/{user_id} to link your Twitter account.",
-            ErrorCode.CLAIM_REQUIRED)
-    
-    market = db.get_market(market_id)
-    if not market:
-        return error_response(404, "Market not found", ErrorCode.MARKET_NOT_FOUND)
-    
-    # Auto-transition: OPEN → RESOLVING when closes_at has passed
-    now = datetime.now(timezone.utc)
-    if market["status"] == MarketStatus.OPEN and market["closes_at"] <= now:
-        db.update_market_status(market_id, MarketStatus.RESOLVING)
-        market["status"] = MarketStatus.RESOLVING
-    
-    if market["status"] != MarketStatus.OPEN:
-        status_msg = "Market is resolving (closed, awaiting resolution)" if market["status"] == MarketStatus.RESOLVING else "Market is not open for trading"
-        return error_response(400, status_msg, ErrorCode.MARKET_CLOSED)
-    
-    if market["closes_at"] <= now:
-        return error_response(400, "Market has closed", ErrorCode.MARKET_CLOSED)
-    
-    # Get user's position
-    position = db.get_position(market_id, user["id"])
-    if not position:
-        return error_response(400, "You have no position in this market", ErrorCode.NO_POSITION)
-    
-    # Check if user has enough shares to sell
-    if req.outcome == Outcome.YES:
-        available_shares = position["yes_shares"]
-    else:
-        available_shares = position["no_shares"]
-    
-    if available_shares < req.shares:
-        return error_response(400,
-            f"Insufficient shares. You have {available_shares:.2f} {req.outcome.value} shares",
-            ErrorCode.INSUFFICIENT_SHARES,
-            detail={"available": available_shares, "requested": req.shares})
-    
-    # Calculate sale using CPMM
-    prob_before = get_cpmm_probability(market["pool"], market["p"])
-    
-    state = CpmmState(pool=market["pool"].copy(), p=market["p"])
-    result = calculate_cpmm_sale(state, req.shares, req.outcome.value)
-    
-    amount_before_fee = result["amount"]
-    if amount_before_fee <= 0:
-        return error_response(400, "Sale would result in zero or negative payout", ErrorCode.ZERO_SHARES)
-    
-    # Apply trade fee (2%)
-    trade_fee = amount_before_fee * TRADE_FEE_RATE
-    amount_after_fee = amount_before_fee - trade_fee
-    
-    prob_after = get_cpmm_probability(result["new_pool"], result["new_p"])
-    
-    # Execute sale
-    # Credit user with the payout (minus fee)
-    db.update_user_balance(user["id"], amount_after_fee)
-    
-    # Pay creator their share of the fee (1%)
-    creator_fee = trade_fee * CREATOR_FEE_SHARE
-    if market["creator_id"] != user["id"]:
-        db.update_user_balance(market["creator_id"], creator_fee)
-    
-    # Update market pool
-    db.update_market_pool(market_id, result["new_pool"], result["new_p"], 0)  # No volume added on sell
-    
-    # Update user's position (reduce shares)
-    db.reduce_position(market_id, user["id"], req.outcome, req.shares)
-    
-    # Invalidate market list cache — probability changed
-    market_cache.invalidate()
-    
-    return SellResponse(
-        market_id=market_id,
-        user_id=user["id"],
-        outcome=req.outcome,
-        shares_sold=req.shares,
-        amount_received=amount_after_fee,
-        fee_paid=trade_fee,
-        probability_before=prob_before,
-        probability_after=prob_after,
-    )
-
-
-@app.get("/markets/{market_id}/positions", response_model=MarketPositions, tags=["trading"])
-async def get_positions(market_id: str):
-    """Get all positions for a market."""
-    _validate_uuid(market_id, "market_id")
-    market = db.get_market(market_id)
-    if not market:
-        return error_response(404, "Market not found", ErrorCode.MARKET_NOT_FOUND)
-    
-    prob = get_cpmm_probability(market["pool"], market["p"])
-    positions = []
-    
-    for pos in db.get_market_positions(market_id):
-        # Calculate current value based on probability
-        current_value = pos["yes_shares"] * prob + pos["no_shares"] * (1 - prob)
-        pnl = current_value - pos["total_invested"]
-        
-        positions.append(Position(
-            user_id=pos["user_id"],
-            market_id=pos["market_id"],
-            yes_shares=pos["yes_shares"],
-            no_shares=pos["no_shares"],
-            total_invested=pos["total_invested"],
-            current_value=current_value,
-            pnl=pnl,
-        ))
-    
-    return MarketPositions(market_id=market_id, positions=positions)
-
-
-@app.get("/markets/{market_id}/history", response_model=MarketHistory, tags=["markets"])
-async def get_market_history(market_id: str):
-    """Get probability history for charts."""
-    _validate_uuid(market_id, "market_id")
-    market = db.get_market(market_id)
-    if not market:
-        return error_response(404, "Market not found", ErrorCode.MARKET_NOT_FOUND)
-    
-    # Get all bets for this market, sorted by time
-    market_bets = sorted(
-        db.get_bets_for_market(market_id),
-        key=lambda x: x["created_at"]
-    )
-    
-    points = []
-    
-    # Add initial point (50% at market creation)
-    points.append(ProbabilityPoint(
-        timestamp=market["created_at"],
-        probability=0.5,
-        volume=0.0,
-    ))
-    
-    # Add point for each bet
-    cumulative_volume = 0.0
-    for bet in market_bets:
-        cumulative_volume += bet["amount"]
-        points.append(ProbabilityPoint(
-            timestamp=bet["created_at"],
-            probability=bet["probability_after"],
-            volume=cumulative_volume,
-        ))
-    
-    return MarketHistory(market_id=market_id, points=points)
-
-
-@app.get("/markets/{market_id}/bets", response_model=List[BetHistoryItem], tags=["trading"])
-async def get_market_bets(market_id: str):
-    """Get all bets for a market."""
-    _validate_uuid(market_id, "market_id")
-    market = db.get_market(market_id)
-    if not market:
-        return error_response(404, "Market not found", ErrorCode.MARKET_NOT_FOUND)
-    
-    # Single JOIN query: bets + usernames (was N+1: 1 + N get_user calls)
-    market_bets = sorted(
-        db.get_bets_for_market_with_users(market_id),
-        key=lambda x: x["created_at"],
-        reverse=True  # Most recent first
-    )
-    
-    items = []
-    for bet in market_bets:
-        items.append(BetHistoryItem(
-            bet_id=bet["id"],
-            user_id=bet["user_id"],
-            username=bet.get("username", "unknown"),
-            outcome=bet["outcome"],
-            amount=bet["amount"],
-            shares=bet["shares"],
-            probability_after=bet["probability_after"],
-            created_at=bet["created_at"],
-        ))
-    
-    return items
-
-
-# =============================================================================
-# Comment Endpoints
-# =============================================================================
-
-@app.get("/markets/{market_id}/comments", response_model=MarketComments, tags=["markets"])
-async def get_comments(market_id: str):
-    """Get all comments for a market."""
-    _validate_uuid(market_id, "market_id")
-    market = db.get_market(market_id)
-    if not market:
-        return error_response(404, "Market not found", ErrorCode.MARKET_NOT_FOUND)
-    
-    raw_comments = db.get_market_comments(market_id)
-    
-    # Build comment tree (top-level + replies)
-    comments_by_id = {}
-    top_level = []
-    
-    for c in raw_comments:
-        comment = Comment(
-            id=c["id"],
-            market_id=c["market_id"],
-            user_id=c["user_id"],
-            username=c.get("username", "unknown"),
-            content=c["content"],
-            created_at=c["created_at"],
-            parent_id=c.get("parent_id"),
-            replies=[],
-        )
-        comments_by_id[c["id"]] = comment
-        
-        if c.get("parent_id") is None:
-            top_level.append(comment)
-    
-    # Attach replies to parents
-    for c in raw_comments:
-        if c.get("parent_id") and c["parent_id"] in comments_by_id:
-            parent = comments_by_id[c["parent_id"]]
-            parent.replies.append(comments_by_id[c["id"]])
-    
-    return MarketComments(
-        market_id=market_id,
-        comments=top_level,
-        total=len(raw_comments),
-    )
-
-
-@app.post("/markets/{market_id}/comments", response_model=Comment, tags=["markets"])
-async def create_comment(market_id: str, req: CommentCreate, user: dict = Depends(require_auth)):
-    """Create a comment on a market."""
-    _validate_uuid(market_id, "market_id")
-    market = db.get_market(market_id)
-    if not market:
-        return error_response(404, "Market not found", ErrorCode.MARKET_NOT_FOUND)
-    
-    # Validate parent comment exists if replying
-    if req.parent_id:
-        parent_comments = db.get_market_comments(market_id)
-        if not any(c["id"] == req.parent_id for c in parent_comments):
-            return error_response(400, "Parent comment not found", ErrorCode.COMMENT_NOT_FOUND)
-    
-    comment_id = str(uuid.uuid4())
-    comment = db.create_comment(
-        comment_id=comment_id,
-        market_id=market_id,
-        user_id=user["id"],
-        content=req.content,
-        parent_id=req.parent_id,
-    )
-    
-    return Comment(
-        id=comment["id"],
-        market_id=comment["market_id"],
-        user_id=comment["user_id"],
-        username=user["username"],
-        content=comment["content"],
-        created_at=comment["created_at"],
-        parent_id=comment.get("parent_id"),
-        replies=[],
-    )
-
-
-# =============================================================================
-# Resolution Committee Endpoints
-# =============================================================================
-
-@app.post("/markets/{market_id}/request-resolution", response_model=ResolutionResult, tags=["markets"])
-async def request_resolution(market_id: str, user: dict = Depends(require_auth)):
-    """
-    Trigger the 9-agent resolution committee to vote on market resolution.
-    
-    Only the market creator can request resolution.
-    The committee will research and vote, with majority (5+) deciding the outcome.
-    """
-    _validate_uuid(market_id, "market_id")
-    market = db.get_market(market_id)
-    if not market:
-        return error_response(404, "Market not found", ErrorCode.MARKET_NOT_FOUND)
-    
-    # Only creator can request resolution
-    if market["creator_id"] != user["id"]:
-        return error_response(403, "Only market creator can request resolution", ErrorCode.FORBIDDEN)
-    
-    if market["status"] == MarketStatus.RESOLVED:
-        return error_response(400, "Market already resolved", ErrorCode.ALREADY_RESOLVED)
-    
-    # Auto-transition OPEN → RESOLVING if closes_at has passed
-    now = datetime.now(timezone.utc)
-    if market["status"] == MarketStatus.OPEN and market["closes_at"] <= now:
-        db.update_market_status(market_id, MarketStatus.RESOLVING)
-        market["status"] = MarketStatus.RESOLVING
-    
-    # Get API keys from environment
-    anthropic_key = os.getenv("ANTHROPIC_API_KEY")
-    brave_key = os.getenv("BRAVE_API_KEY")
-    
-    if not anthropic_key or not brave_key:
-        return error_response(500, "Resolution service not configured", ErrorCode.SERVICE_UNAVAILABLE)
-    
-    # Run the resolution committee
-    status, outcome, votes = await resolver_resolve_market(
-        market_id=market_id,
-        market_title=market["title"],
-        market_description=market.get("description", ""),
-        resolution_criteria=market.get("description", ""),  # Use description as criteria
-        anthropic_key=anthropic_key,
-        brave_key=brave_key,
-    )
-    
-    # Save votes
-    vote_dicts = [
-        {
-            "agent_id": v.agent_id,
-            "vote": v.vote,
-            "reasoning": v.reasoning,
-            "sources": v.sources,
-            "created_at": v.created_at.isoformat(),
-        }
-        for v in votes
-    ]
-    db.save_resolution_votes(market_id, vote_dicts)
-    
-    # If resolved, update the market
-    resolved_at = None
-    if status == "resolved" and outcome:
-        outcome_enum = Outcome.YES if outcome == "YES" else Outcome.NO
-        db.resolve_market(market_id, outcome_enum)
-        _calculate_and_distribute_payouts(market_id, outcome_enum)
-        
-        resolved_at = datetime.now(timezone.utc)
-    
-    # Build response
-    return ResolutionResult(
-        market_id=market_id,
-        status=status,
-        outcome=Outcome(outcome) if outcome else None,
-        votes_yes=sum(1 for v in votes if v.vote == "YES"),
-        votes_no=sum(1 for v in votes if v.vote == "NO"),
-        total_votes=len(votes),
-        votes=[
-            ResolutionVote(
-                agent_id=v.agent_id,
-                vote=Outcome(v.vote),
-                reasoning=v.reasoning,
-                sources=v.sources,
-                created_at=v.created_at,
-            )
-            for v in votes
-        ],
-        resolved_at=resolved_at,
-    )
-
-
-@app.get("/markets/{market_id}/resolution-votes", response_model=ResolutionResult, tags=["markets"])
-async def get_resolution_votes(market_id: str):
-    """Get the resolution committee votes for a market."""
-    _validate_uuid(market_id, "market_id")
-    market = db.get_market(market_id)
-    if not market:
-        return error_response(404, "Market not found", ErrorCode.MARKET_NOT_FOUND)
-    
-    votes = db.get_resolution_votes(market_id)
-    
-    if not votes:
-        return error_response(404, "No resolution votes found for this market", ErrorCode.NOT_FOUND)
-    
-    yes_votes = sum(1 for v in votes if v["vote"] == "YES")
-    no_votes = sum(1 for v in votes if v["vote"] == "NO")
-    
-    # Determine status
-    if market["status"] == MarketStatus.RESOLVED:
-        status = "resolved"
-        outcome = market["resolution"]
-    elif yes_votes >= 5:
-        status = "resolved"
-        outcome = Outcome.YES
-    elif no_votes >= 5:
-        status = "resolved"
-        outcome = Outcome.NO
-    else:
-        status = "disputed"
-        outcome = None
-    
-    return ResolutionResult(
-        market_id=market_id,
-        status=status,
-        outcome=outcome,
-        votes_yes=yes_votes,
-        votes_no=no_votes,
-        total_votes=len(votes),
-        votes=[
-            ResolutionVote(
-                agent_id=v["agent_id"],
-                vote=Outcome(v["vote"]),
-                reasoning=v["reasoning"],
-                sources=v.get("sources", []),
-                created_at=datetime.fromisoformat(v["created_at"]) if isinstance(v["created_at"], str) else v["created_at"],
-            )
-            for v in votes
-        ],
-        resolved_at=market.get("resolved_at"),
-    )
-
-
-# =============================================================================
-# User Endpoints
-# =============================================================================
-
-@app.get("/me", response_model=UserMe, tags=["agents"])
-async def get_me(user: dict = Depends(require_auth)):
-    """Get current user profile with balance."""
-    return UserMe(
-        id=user["id"],
-        username=user["username"],
-        display_name=user["display_name"],
-        balance=user["balance"],
-        created_at=user["created_at"],
-        markets_created=user["markets_created"],
-        total_bets=user["total_bets"],
-        profit_all_time=user["profit_all_time"],
-    )
-
-
-@app.get("/me/positions", response_model=PortfolioResponse, tags=["agents"])
-async def get_my_positions(user: dict = Depends(require_auth)):
-    """Get all positions for the authenticated agent across all markets.
-
-    Returns a portfolio overview with per-market positions and an aggregate summary.
-    Saves agents from making N+1 calls to /markets/{id}/positions.
-    """
-    positions = db.get_user_positions(user["id"])
-    items: List[PortfolioPosition] = []
-    total_invested = 0.0
-    total_current_value = 0.0
-    open_count = 0
-    resolved_count = 0
-
-    for pos in positions:
-        market = db.get_market(pos["market_id"])
-        if not market:
-            continue
-
-        prob = get_cpmm_probability(market["pool"], market["p"])
-        current_value = pos["yes_shares"] * prob + pos["no_shares"] * (1 - prob)
-        pnl = current_value - pos["total_invested"]
-
-        is_open = market["status"] in (MarketStatus.OPEN, MarketStatus.RESOLVING)
-        if is_open:
-            open_count += 1
-        else:
-            resolved_count += 1
-
-        total_invested += pos["total_invested"]
-        total_current_value += current_value
-
-        items.append(PortfolioPosition(
-            market_id=pos["market_id"],
-            market_title=market["title"],
-            market_status=market["status"],
-            yes_shares=pos["yes_shares"],
-            no_shares=pos["no_shares"],
-            total_invested=pos["total_invested"],
-            current_value=round(current_value, 4),
-            pnl=round(pnl, 4),
-            current_probability=round(prob, 6),
-        ))
-
-    return PortfolioResponse(
-        positions=items,
-        summary=PortfolioSummary(
-            total_invested=round(total_invested, 4),
-            total_current_value=round(total_current_value, 4),
-            total_pnl=round(total_current_value - total_invested, 4),
-            open_positions=open_count,
-            resolved_positions=resolved_count,
-        ),
-    )
-
-
-@app.get("/me/bets", response_model=List[UserBetHistoryItem], tags=["agents"])
-async def get_my_bets(
-    limit: int = 50,
-    offset: int = 0,
-    user: dict = Depends(require_auth),
-):
-    """Get the authenticated agent's trade history across all markets.
-
-    Supports basic pagination via limit/offset.
-    Returns bets sorted by created_at DESC (newest first).
-    """
-    if limit < 1:
-        limit = 1
-    if limit > 200:
-        limit = 200
-    if offset < 0:
-        offset = 0
-
-    all_bets = db.get_bets_for_user(user["id"])
-    # Sort newest first
-    all_bets.sort(key=lambda b: b["created_at"], reverse=True)
-    page = all_bets[offset : offset + limit]
-
-    items: List[UserBetHistoryItem] = []
-    for bet in page:
-        market = db.get_market(bet["market_id"])
-        market_title = market["title"] if market else "Unknown market"
-        items.append(UserBetHistoryItem(
-            bet_id=bet["id"],
-            market_id=bet["market_id"],
-            market_title=market_title,
-            outcome=bet["outcome"],
-            amount=bet["amount"],
-            shares=bet["shares"],
-            avg_price=bet["avg_price"],
-            probability_before=bet["probability_before"],
-            probability_after=bet["probability_after"],
-            created_at=bet["created_at"],
-        ))
-
-    return items
-
-
-@app.get("/users/{user_id}", response_model=UserProfile, tags=["agents"])
-async def get_user(user_id: str):
-    """Get public user profile."""
-    _validate_uuid(user_id, "user_id")
-    user = db.get_user(user_id)
-    if not user:
-        return error_response(404, "User not found", ErrorCode.USER_NOT_FOUND)
-    
-    return UserProfile(
-        id=user["id"],
-        username=user["username"],
-        display_name=user["display_name"],
-        balance=user["balance"],
-        created_at=user["created_at"],
-        markets_created=user["markets_created"],
-        total_bets=user["total_bets"],
-        profit_all_time=user["profit_all_time"],
-        twitter_handle=user.get("twitter_handle"),
-    )
-
-
-# =============================================================================
-# Agent Reputation
-# =============================================================================
-
-@app.get("/agents/{agent_id}/reputation", response_model=AgentReputationResponse, tags=["agents"])
-async def get_agent_reputation(agent_id: str):
-    """
-    Get the multi-dimensional reputation profile for an agent.
-    
-    Reputation is computed on-the-fly from trading history, resolution votes,
-    market creation quality, and participation. All scores are 0-100.
-    
-    Dimensions:
-    - **trading**: P&L performance, win rate, volume
-    - **resolution**: Accuracy of resolution committee votes  
-    - **creation**: Quality of markets created (volume, bet count)
-    - **participation**: Overall platform engagement
-    
-    The overall_score is a weighted composite, mapped to a tier:
-    New (<40) → Bronze (40-54) → Silver (55-69) → Gold (70-84) → Platinum (85+)
-    """
-    user = db.get_user(agent_id)
-    if not user:
-        # Also try by username
-        user = db.get_user_by_username(agent_id)
-    if not user:
-        return error_response(404, "Agent not found", ErrorCode.AGENT_NOT_FOUND)
-    
-    # Gather all data needed for reputation calculation.
-    # Optimized (#54): targeted queries instead of full-table loads.
-    # Previously: db.markets (ALL markets) + db.bets (ALL bets) → O(N) each.
-    # Now: fetch only the markets/bets relevant to this user → O(K) where K << N.
-    user_bets = db.get_bets_for_user(user["id"])
-
-    # 1. Collect the market IDs we actually need:
-    #    - Markets the user bet on (for trading score)
-    #    - Markets created by the user (for creation score)
-    bet_market_ids = {b["market_id"] for b in user_bets}
-    user_created_markets = db.get_markets_by_creator(user["id"])
-    created_market_ids = {m["id"] for m in user_created_markets}
-
-    # 2. Batch-fetch resolution votes + comment count
-    rep_data = db.get_reputation_data(user["id"])
-    all_resolution_votes = rep_data["resolution_votes"]
-    comments_count = rep_data["comments_count"]
-
-    # Include market IDs from this user's resolution votes
-    vote_market_ids = {
-        v.get("market_id") for v in all_resolution_votes
-        if v.get("agent_id") == user["id"] and v.get("market_id")
-    }
-
-    # 3. Fetch only the markets we need (union of all relevant IDs)
-    all_needed_ids = bet_market_ids | created_market_ids | vote_market_ids
-    markets = db.get_markets_by_ids(all_needed_ids)
-
-    # 4. Fetch only bets on user's created markets (for creation score)
-    all_bets = db.get_bets_on_markets(created_market_ids) if created_market_ids else []
-
-    # Compute reputation (pure function — no DB access)
-    rep = compute_reputation(
-        user=user,
-        user_bets=user_bets,
-        markets=markets,
-        all_bets=all_bets,
-        resolution_votes=all_resolution_votes,
-        comments_count=comments_count,
-    )
-    
-    return AgentReputationResponse(
-        agent_id=rep.agent_id,
-        username=rep.username,
-        overall_score=rep.overall_score,
-        tier=rep.tier,
-        trading=TradingScoreResponse(
-            score=rep.trading.score,
-            total_pnl=rep.trading.total_pnl,
-            resolved_bets=rep.trading.resolved_bets,
-            win_rate=rep.trading.win_rate,
-            total_volume=rep.trading.total_volume,
-        ),
-        resolution=ResolutionScoreResponse(
-            score=rep.resolution.score,
-            total_votes=rep.resolution.total_votes,
-            correct_votes=rep.resolution.correct_votes,
-            accuracy=rep.resolution.accuracy,
-        ),
-        creation=CreationScoreResponse(
-            score=rep.creation.score,
-            markets_created=rep.creation.markets_created,
-            total_volume_attracted=rep.creation.total_volume_attracted,
-            total_bets_attracted=rep.creation.total_bets_attracted,
-            avg_volume_per_market=rep.creation.avg_volume_per_market,
-            avg_bets_per_market=rep.creation.avg_bets_per_market,
-            resolved_cleanly=rep.creation.resolved_cleanly,
-            disputed=rep.creation.disputed,
-        ),
-        participation=ParticipationScoreResponse(
-            score=rep.participation.score,
-            total_bets=rep.participation.total_bets,
-            markets_traded_in=rep.participation.markets_traded_in,
-            markets_created=rep.participation.markets_created,
-            comments_count=rep.participation.comments_count,
-        ),
-    )
-
-
-# =============================================================================
-# Agent Registration
-# =============================================================================
-
-@app.post("/agents/register", response_model=AgentRegisteredWithClaim, tags=["agents"])
-async def register_agent(req: AgentRegister, request: Request, response: Response):
-    """
-    Register a new agent and get an API key.
-    
-    The API key is only returned ONCE — save it securely!
-    
-    Returns a verification_code and claim_url for human-agent linking.
-    The human must tweet the verification code to claim the agent.
-    
-    Rate limited: max 5 registrations per IP per hour.
-    """
-    # ── Rate limit: registrations per IP ──
-    client_ip = request.client.host if request.client else "unknown"
-    allowed, info = rate_limiter.check(
-        f"register:{client_ip}",
-        max_requests=MAX_REGISTRATIONS_PER_HOUR,
-        window_seconds=3600,
-    )
-    if not allowed:
-        raise_rate_limited(
-            f"Registration rate limit exceeded ({MAX_REGISTRATIONS_PER_HOUR}/hour per IP). {info['detail']}",
-            info,
-        )
-    set_rate_limit_headers(response, info)
-
-    # Normalize username to lowercase (case-insensitive uniqueness)
-    username = req.username.lower()
-    
-    # Check if username is taken
-    if db.get_user_by_username(username):
-        return error_response(400, "Username already taken", ErrorCode.ALREADY_EXISTS)
-    
-    # Generate API key, user ID, and verification code
-    api_key = generate_api_key()
-    user_id = str(uuid.uuid4())
-    verification_code = generate_verification_code()
-    
-    # Create user with hashed API key
-    user = db.create_user(
-        user_id=user_id,
-        username=username,
-        balance=STARTING_BALANCE,
-        api_key_hash=hash_api_key(api_key),
-        description=req.description or "",
-        status="pending",
-        verification_code=verification_code,
-    )
-    
-    # Update display name if provided
-    if req.display_name:
-        db.update_user_display_name(user_id, req.display_name)
-        user["display_name"] = req.display_name
-    
-    return AgentRegisteredWithClaim(
-        user_id=user["id"],
-        username=user["username"],
-        display_name=user["display_name"],
-        description=user.get("description", ""),
-        api_key=api_key,  # Only time we return the raw key!
-        balance=user["balance"],
-        created_at=user["created_at"],
-        markets_created=user.get("markets_created", 0),
-        total_bets=user.get("total_bets", 0),
-        profit_all_time=user.get("profit_all_time", 0.0),
-        status=AgentStatus.PENDING,
-        verification_code=verification_code,
-        claim_url=f"/claim/{user_id}",
-    )
-
-
-@app.post("/agents/reset-key", response_model=AgentKeyReset, tags=["agents"])
-async def reset_api_key(user: dict = Depends(require_auth)):
-    """
-    Reset your API key. Requires current valid API key.
-    
-    Old key becomes invalid immediately.
-    """
-    new_key = generate_api_key()
-    db.update_api_key(user["id"], hash_api_key(new_key))
-    
-    return AgentKeyReset(
-        user_id=user["id"],
-        api_key=new_key,
-    )
 
 
 # Admin secret for privileged operations (MUST be set via ADMIN_SECRET env var)
@@ -1626,358 +453,8 @@ if not ADMIN_SECRET:
     print("WARNING: ADMIN_SECRET not set — admin endpoints will be disabled")
 
 
-@app.delete("/admin/users/{username}", tags=["admin"])
-async def admin_delete_user(username: str, request: Request, x_admin_secret: str = Header(None)):
-    """
-    Delete a user by username (admin only).
-    Requires X-Admin-Secret header.
-    """
-    if not ADMIN_SECRET:
-        return error_response(503, "Admin endpoints disabled — ADMIN_SECRET not configured", ErrorCode.SERVICE_UNAVAILABLE)
-    
-    # Rate limit admin endpoints to mitigate brute-force attacks
-    client_ip = request.client.host if request.client else "unknown"
-    allowed, info = rate_limiter.check(f"admin:{client_ip}", max_requests=10, window_seconds=60)
-    if not allowed:
-        return error_response(429, f"Admin rate limit exceeded. {info['detail']}", ErrorCode.RATE_LIMITED)
-    
-    # Use constant-time comparison to prevent timing attacks (see #55)
-    if not secrets.compare_digest(x_admin_secret or "", ADMIN_SECRET):
-        return error_response(403, "Invalid admin secret", ErrorCode.FORBIDDEN)
-    
-    user = db.get_user_by_username(username)
-    if not user:
-        return error_response(404, f"User '{username}' not found", ErrorCode.USER_NOT_FOUND)
-    
-    # Delete user (we need to add this method)
-    db.delete_user(user["id"])
-    
-    return {"deleted": True, "username": username, "user_id": user["id"]}
-
-
-@app.post("/admin/users/{username}/regenerate-key", tags=["admin"])
-async def admin_regenerate_api_key(username: str, request: Request, x_admin_secret: str = Header(None)):
-    """
-    Regenerate API key for a user (admin only).
-    Returns the new API key — save it, it won't be shown again!
-    """
-    if not ADMIN_SECRET:
-        return error_response(503, "Admin endpoints disabled — ADMIN_SECRET not configured", ErrorCode.SERVICE_UNAVAILABLE)
-    
-    # Rate limit admin endpoints to mitigate brute-force attacks
-    client_ip = request.client.host if request.client else "unknown"
-    allowed, info = rate_limiter.check(f"admin:{client_ip}", max_requests=10, window_seconds=60)
-    if not allowed:
-        return error_response(429, f"Admin rate limit exceeded. {info['detail']}", ErrorCode.RATE_LIMITED)
-    
-    # Use constant-time comparison to prevent timing attacks (see #55)
-    if not secrets.compare_digest(x_admin_secret or "", ADMIN_SECRET):
-        return error_response(403, "Invalid admin secret", ErrorCode.FORBIDDEN)
-    
-    user = db.get_user_by_username(username)
-    if not user:
-        return error_response(404, f"User '{username}' not found", ErrorCode.USER_NOT_FOUND)
-    
-    # Generate new API key
-    new_api_key = generate_api_key()
-    key_hash = hash_api_key(new_api_key)
-    
-    # Update in database
-    db.update_user_api_key(user["id"], key_hash)
-    
-    return {
-        "username": username,
-        "user_id": user["id"],
-        "api_key": new_api_key,
-        "warning": "Save this key! It will not be shown again."
-    }
-
-
-@app.get("/claim/{user_id}", response_model=ClaimPageInfo, tags=["agents"])
-async def get_claim_info(user_id: str):
-    """
-    Get claim page info for an agent (public, no auth required).
-    
-    Returns the verification code and instructions for claiming.
-    """
-    _validate_uuid(user_id, "user_id")
-    user = db.get_user(user_id)
-    if not user:
-        return error_response(404, "Agent not found", ErrorCode.AGENT_NOT_FOUND)
-    
-    if not user.get("verification_code"):
-        return error_response(400, "Agent has no verification code", ErrorCode.INVALID_INPUT)
-    
-    if user.get("status") == "claimed":
-        return error_response(400, "Agent already claimed", ErrorCode.ALREADY_CLAIMED)
-    
-    instructions = (
-        f"To claim this agent, post a tweet containing the verification code: {user['verification_code']}\n\n"
-        f"Example tweet: 'I'm claiming my MoltMarkets agent! Verification: {user['verification_code']}'\n\n"
-        f"After posting, submit the tweet URL to complete the claim."
-    )
-    
-    return ClaimPageInfo(
-        user_id=user["id"],
-        username=user["username"],
-        display_name=user["display_name"],
-        verification_code=user["verification_code"],
-        instructions=instructions,
-    )
-
-
-@app.post("/agents/claim", response_model=ClaimResponse, tags=["agents"])
-async def claim_agent(req: ClaimRequest):
-    """
-    Claim an agent by providing a tweet URL with the verification code.
-    
-    The tweet must contain the agent's verification code.
-    """
-    # Get the user/agent
-    user = db.get_user(req.user_id)
-    if not user:
-        return error_response(404, "Agent not found", ErrorCode.AGENT_NOT_FOUND)
-    
-    if user.get("status") == "claimed":
-        return error_response(400, "Agent already claimed", ErrorCode.ALREADY_CLAIMED)
-    
-    if not user.get("verification_code"):
-        return error_response(400, "Agent has no verification code", ErrorCode.INVALID_INPUT)
-    
-    # Validate tweet URL format
-    if not is_valid_twitter_url(req.tweet_url):
-        return error_response(400,
-            "Invalid tweet URL. Must be a twitter.com or x.com status URL (e.g., https://twitter.com/user/status/123456)",
-            ErrorCode.INVALID_INPUT)
-    
-    # Extract tweet ID from URL
-    tweet_id = extract_tweet_id(req.tweet_url)
-    if not tweet_id:
-        return error_response(400, "Could not extract tweet ID from URL", ErrorCode.INVALID_INPUT)
-    
-    # Fetch the tweet content
-    tweet_data = await fetch_tweet(tweet_id)
-    
-    # Verify the tweet contains the verification code
-    tweet_text = tweet_data.get("text", "")
-    if not verify_tweet_contains_code(tweet_text, user["verification_code"]):
-        return error_response(400,
-            f"Verification failed: tweet does not contain the code '{user['verification_code']}'. "
-            f"Please ensure your tweet includes the exact verification code.",
-            ErrorCode.VERIFICATION_FAILED)
-    
-    # Extract twitter handle from the tweet URL
-    twitter_handle = extract_twitter_handle(req.tweet_url)
-    
-    # Mark agent as claimed and store twitter handle
-    db.update_user_status(req.user_id, "claimed")
-    if twitter_handle:
-        db.update_user_twitter_handle(req.user_id, twitter_handle)
-    
-    return ClaimResponse(
-        success=True,
-        message=f"Agent '{user['username']}' successfully claimed!",
-        user_id=user["id"],
-        username=user["username"],
-        display_name=user["display_name"],
-        status=AgentStatus.CLAIMED,
-    )
-
-
 # =============================================================================
-# Human Registration
-# =============================================================================
-
-@app.post("/humans/register", response_model=HumanRegistered, tags=["agents"])
-async def register_human(req: HumanRegister, request: Request, response: Response):
-    """
-    Register a human user for chat.
-    
-    Lightweight registration — no Twitter verification needed.
-    Human users can post in the 'humans' chat channel.
-    The API key is only returned ONCE — save it!
-    
-    Rate limited: max 5 registrations per IP per hour.
-    """
-    # Rate limit: registrations per IP
-    client_ip = request.client.host if request.client else "unknown"
-    allowed, info = rate_limiter.check(
-        f"register-human:{client_ip}",
-        max_requests=MAX_REGISTRATIONS_PER_HOUR,
-        window_seconds=3600,
-    )
-    if not allowed:
-        raise_rate_limited(
-            f"Registration rate limit exceeded ({MAX_REGISTRATIONS_PER_HOUR}/hour per IP). {info['detail']}",
-            info,
-        )
-    set_rate_limit_headers(response, info)
-
-    # Normalize username to lowercase
-    username = req.username.lower()
-    
-    # Check if username is taken (shared namespace with agents)
-    if db.get_user_by_username(username):
-        return error_response(400, "Username already taken", ErrorCode.ALREADY_EXISTS)
-    
-    # Generate API key and user ID
-    api_key = generate_api_key()
-    user_id = str(uuid.uuid4())
-    
-    # Create human user (no verification code, status='claimed' immediately, user_type='human')
-    user = db.create_user(
-        user_id=user_id,
-        username=username,
-        balance=STARTING_BALANCE,
-        api_key_hash=hash_api_key(api_key),
-        description="",
-        status="claimed",  # Humans are immediately active
-        verification_code=None,
-        user_type="human",
-    )
-    
-    # Update display name if provided
-    if req.display_name:
-        db.update_user_display_name(user_id, req.display_name)
-        user["display_name"] = req.display_name
-    
-    return HumanRegistered(
-        user_id=user["id"],
-        username=user["username"],
-        display_name=user["display_name"],
-        api_key=api_key,
-        balance=user["balance"],
-        user_type="human",
-        created_at=user["created_at"],
-        markets_created=user.get("markets_created", 0),
-        total_bets=user.get("total_bets", 0),
-        profit_all_time=user.get("profit_all_time", 0.0),
-    )
-
-
-# =============================================================================
-# Leaderboard
-# =============================================================================
-
-@app.get("/leaderboard", response_model=List[LeaderboardEntry], tags=["agents"])
-async def get_leaderboard():
-    """Get leaderboard sorted by profit (only shows claimed/verified agents)."""
-    # Single aggregate query with CTEs (was: 3 full-table loads + O(users × bets) iteration)
-    leaderboard_data = db.get_leaderboard_data()
-    
-    return [
-        LeaderboardEntry(
-            user_id=entry["user_id"],
-            username=entry["username"],
-            pnl=entry["pnl"],
-            total_volume=entry["total_volume"],
-            win_rate=entry["win_rate"],
-        )
-        for entry in leaderboard_data
-    ]
-
-
-# =============================================================================
-# Chat Endpoints
-# =============================================================================
-
-@app.post("/chat", response_model=ChatMessage, tags=["chat"])
-async def send_chat_message(req: ChatMessageCreate, response: Response, channel: str = "agents", user: dict = Depends(require_auth)):
-    """
-    Send a chat message.
-    
-    Auth required. Max 500 characters. Rate limited: 10 messages per minute per user.
-    
-    Query params:
-        channel: Chat channel to post in ('agents' or 'humans', default 'agents').
-                 The 'humans' channel is restricted to human users only (user_type='human').
-                 Agents (user_type='agent') will get a 403 if they try to post in 'humans'.
-    """
-    # Validate channel
-    if channel not in ("agents", "humans"):
-        return error_response(400, "Invalid channel. Must be 'agents' or 'humans'.", ErrorCode.INVALID_INPUT)
-    
-    # Enforce humans-only restriction
-    if channel == "humans" and user.get("user_type", "agent") == "agent":
-        return error_response(403,
-            "Only human users can post in the 'humans' channel. Agents can read but not write.",
-            ErrorCode.FORBIDDEN)
-    
-    # Rate limit: chat messages per user
-    allowed, info = rate_limiter.check(
-        f"chat:{user['id']}",
-        max_requests=MAX_CHAT_MESSAGES_PER_MINUTE,
-        window_seconds=60,
-    )
-    if not allowed:
-        raise_rate_limited(
-            f"Chat rate limit exceeded ({MAX_CHAT_MESSAGES_PER_MINUTE}/minute). {info['detail']}",
-            info,
-        )
-    set_rate_limit_headers(response, info)
-    
-    message = db.create_chat_message(
-        user_id=user["id"],
-        username=user["username"],
-        text=req.text,
-        channel=channel,
-    )
-    
-    return ChatMessage(
-        id=message["id"],
-        username=message["username"],
-        text=message["text"],
-        channel=message.get("channel", "agents"),
-        created_at=message["created_at"],
-    )
-
-
-@app.get("/chat", response_model=List[ChatMessage], tags=["chat"])
-async def get_chat_messages(limit: int = 50, since: Optional[str] = None, channel: str = "agents"):
-    """
-    Get recent chat messages.
-    
-    Query params:
-        limit: Number of messages to return (default 50, max 200).
-        since: ISO 8601 timestamp — only return messages after this time (for polling).
-        channel: Chat channel to read from ('agents' or 'humans', default 'agents').
-    
-    Returns messages sorted by created_at DESC (newest first).
-    """
-    # Validate channel
-    if channel not in ("agents", "humans"):
-        return error_response(400, "Invalid channel. Must be 'agents' or 'humans'.", ErrorCode.INVALID_INPUT)
-    
-    # Clamp limit
-    if limit < 1:
-        limit = 1
-    if limit > 200:
-        limit = 200
-    
-    # Parse since parameter
-    since_dt = None
-    if since:
-        try:
-            since_dt = datetime.fromisoformat(since.replace('Z', '+00:00'))
-        except ValueError:
-            return error_response(400, "Invalid 'since' parameter. Use ISO 8601 format (e.g. 2026-01-30T23:00:00Z).", ErrorCode.INVALID_INPUT)
-    
-    messages = db.get_chat_messages(limit=limit, since=since_dt, channel=channel)
-    
-    return [
-        ChatMessage(
-            id=str(m["id"]),
-            username=m["username"],
-            text=m["text"],
-            channel=m.get("channel", "agents"),
-            created_at=m["created_at"],
-        )
-        for m in messages
-    ]
-
-
-# =============================================================================
-# Agent Discovery — /skill.md
+# Agent Discovery — /skill.md (string template used by routes/meta.py)
 # =============================================================================
 
 _SKILL_MD = f"""\
@@ -2163,11 +640,11 @@ Rate limit headers are returned on relevant responses:
 All errors return JSON with a machine-readable error code:
 
 ```json
-{{
+{{{{
   "error": "Human-readable error message",
   "code": "ERROR_CODE",
-  "detail": {{}}
-}}
+  "detail": {{{{}}}}
+}}}}
 ```
 
 The `detail` field is optional and provides structured context (e.g., balance, retry timing).
@@ -2179,66 +656,34 @@ Common status codes: `400` (bad request), `401` (auth required), `403` (forbidde
 """
 
 
-@app.get("/skill.md", tags=["meta"], include_in_schema=False)
-async def get_skill_md():
-    """Return a markdown skill file describing this API for agent auto-discovery."""
-    from fastapi.responses import PlainTextResponse
-    return PlainTextResponse(_SKILL_MD, media_type="text/markdown; charset=utf-8")
-
-
 # =============================================================================
-# Health Check
+# Route Registration
 # =============================================================================
-
-@app.get("/health", tags=["meta"])
-async def health():
-    from fastapi.responses import JSONResponse
-
-    # Verify database is reachable
-    db_status = "ok"
-    try:
-        conn = db._get_conn()
-        try:
-            with conn.cursor() as cur:
-                cur.execute("SELECT 1")
-        finally:
-            db._put_conn(conn)
-    except Exception:
-        db_status = "unreachable"
-
-    if db_status != "ok":
-        return JSONResponse(
-            status_code=503,
-            content={
-                "status": "degraded",
-                "db": db_status,
-            },
-        )
-
-    market_count = db.count_markets()
-    user_count = db.count_users()
-    return {
-        "status": "ok",
-        "db": "ok",
-        "markets": market_count,
-        "users": user_count,
-        "idempotency_keys_cached": idempotency_store.size,
-        "currency": {
-            "symbol": CURRENCY_SYMBOL,
-            "name": CURRENCY_NAME,
-        },
-    }
+# Route handlers have been extracted to routes/ modules for maintainability.
+# Each module defines an APIRouter; we include them here.
+#
+# Imports are deferred inside a function to avoid circular imports —
+# route modules do `from api import db, ...` which requires api.py's
+# module-level symbols to be defined first.
 
 
-@app.get("/currency", tags=["meta"])
-async def get_currency():
-    """Get platform currency info. MoltMarkets uses points (ŧ), not real money."""
-    return {
-        "symbol": CURRENCY_SYMBOL,
-        "name": CURRENCY_NAME,
-        "starting_balance": STARTING_BALANCE,
-        "note": "MoltMarkets uses points (ŧ), not real money. All balances and amounts are in points.",
-    }
+def _register_routes():
+    from routes.markets import router as markets_router
+    from routes.trading import router as trading_router
+    from routes.agents import router as agents_router
+    from routes.chat import router as chat_router
+    from routes.admin import router as admin_router
+    from routes.meta import router as meta_router
+
+    app.include_router(markets_router)
+    app.include_router(trading_router)
+    app.include_router(agents_router)
+    app.include_router(chat_router)
+    app.include_router(admin_router)
+    app.include_router(meta_router)
+
+
+_register_routes()
 
 
 # =============================================================================
@@ -2247,23 +692,20 @@ async def get_currency():
 
 if __name__ == "__main__":
     import asyncio
-    
+
     print("=" * 60)
     print("MoltMarkets API - Quick Test")
     print("=" * 60)
-    
+
     async def run_test():
-        # Simulate requests without running actual server
-        
-        # 1. Create a user
+        from datetime import timedelta
+
         print("\n1. Creating demo user...")
         db.create_user("test-user", "test_user", balance=STARTING_BALANCE)
         user = db.get_user("test-user")
         print(f"   User: {user['username']}, Balance: {user['balance']}ŧ")
-        
-        # 2. Create a market
+
         print("\n2. Creating a market...")
-        from datetime import timedelta
         market_id = str(uuid.uuid4())
         market = db.create_market(
             market_id=market_id,
@@ -2276,23 +718,21 @@ if __name__ == "__main__":
         prob = get_cpmm_probability(market["pool"], market["p"])
         print(f"   Market: {market['title'][:40]}...")
         print(f"   Pool: {market['pool']}, Probability: {prob:.2%}")
-        
-        # 3. Place a YES bet
+
         print("\n3. Placing 50ŧ bet on YES...")
         state = CpmmState(pool=market["pool"].copy(), p=market["p"])
         result = calculate_cpmm_purchase(state, 50, "YES")
-        
+
         shares = result["shares"]
         db.update_user_balance("test-user", -50)
         db.update_market_pool(market_id, result["new_pool"], result["new_p"], 50)
         db.update_position(market_id, "test-user", Outcome.YES, shares, 50)
-        
+
         new_prob = get_cpmm_probability(result["new_pool"], result["new_p"])
         print(f"   Shares received: {shares:.2f}")
         print(f"   Avg price: {50/shares:.4f}ŧ per share")
         print(f"   Probability: {prob:.2%} → {new_prob:.2%}")
-        
-        # 4. Check position
+
         print("\n4. Checking position...")
         pos = db.get_position(market_id, "test-user")
         current_value = pos["yes_shares"] * new_prob
@@ -2300,15 +740,14 @@ if __name__ == "__main__":
         print(f"   Total invested: {pos['total_invested']:.2f}ŧ")
         print(f"   Current value: {current_value:.2f}ŧ")
         print(f"   P&L: {current_value - pos['total_invested']:.2f}ŧ")
-        
-        # 5. Check user balance
+
         print("\n5. Checking user balance...")
         user = db.get_user("test-user")
         print(f"   Balance: {user['balance']:.2f}ŧ")
         print(f"   Total bets: {user['total_bets']}")
-        
+
         print("\n" + "=" * 60)
         print("Test complete! Run with: uvicorn api:app --reload")
         print("=" * 60)
-    
+
     asyncio.run(run_test())
